Move functions between Client and LocalNode as appropriate (#2847)
* Move functions from LocalNode to RemoteNode that only use the latter.

* Move methods from LocalNode to Client.

* Remove redundant local_query.

* Use only one local node in assign_new_chain_to_key.
