+++ linera-core/src/local_node.rs
@@ -4,13 +4,11 @@
 
 use std::{
     collections::{HashSet, VecDeque},
-    future::Future,
     sync::Arc,
 };
 
-use futures::{future, stream::FuturesUnordered};
 use linera_base::{
-    data_types::{ArithmeticError, Blob, BlockHeight, UserApplicationDescription},
+    data_types::{ArithmeticError, Blob, UserApplicationDescription},
     identifiers::{BlobId, ChainId, MessageId, UserApplicationId},
 };
 use linera_chain::{
@@ -22,16 +20,14 @@ use linera_chain::{
 use linera_execution::{ExecutionError, Query, Response, SystemExecutionError};
 use linera_storage::Storage;
 use linera_views::views::ViewError;
-use rand::{prelude::SliceRandom, thread_rng};
 use thiserror::Error;
 use tokio::sync::OwnedRwLockReadGuard;
 use tracing::{instrument, warn};
 
 use crate::{
     data_types::{BlockHeightRange, ChainInfo, ChainInfoQuery, ChainInfoResponse},
-    node::{NodeError, ValidatorNode},
+    node::NodeError,
     notifier::Notifier,
-    remote_node::RemoteNode,
     worker::{WorkerError, WorkerState},
 };
 
@@ -64,15 +60,6 @@ pub enum LocalNodeError {
     #[error("Local node operation failed: {0}")]
     WorkerError(#[from] WorkerError),
 
-    #[error(
-        "Failed to download certificates and update local node to the next height \
-         {target_next_block_height} of chain {chain_id:?}"
-    )]
-    CannotDownloadCertificates {
-        chain_id: ChainId,
-        target_next_block_height: BlockHeight,
-    },
-
     #[error("Failed to read blob {blob_id:?} of chain {chain_id:?}")]
     CannotReadLocalBlob { chain_id: ChainId, blob_id: BlobId },
 
@@ -81,9 +68,6 @@ pub enum LocalNodeError {
 
     #[error("The chain info response received from the local node is invalid")]
     InvalidChainInfoResponse,
-
-    #[error(transparent)]
-    NodeError(#[from] NodeError),
 }
 
 impl LocalNodeError {
@@ -107,10 +91,6 @@ impl LocalNodeError {
             LocalNodeError::WorkerError(WorkerError::BlobsNotFound(blob_ids)) => {
                 Some(blob_ids.clone())
             }
-            LocalNodeError::NodeError(NodeError::BlobNotFoundOnRead(blob_id)) => {
-                Some(vec![*blob_id])
-            }
-            LocalNodeError::NodeError(NodeError::BlobsNotFound(blob_ids)) => Some(blob_ids.clone()),
             _ => None,
         }
     }
@@ -271,55 +251,6 @@ where
         Ok(found_blobs)
     }
 
-    #[instrument(level = "trace", skip_all)]
-    pub async fn try_process_certificates(
-        &self,
-        remote_node: &RemoteNode<impl ValidatorNode>,
-        chain_id: ChainId,
-        certificates: Vec<Certificate>,
-        notifier: &impl Notifier,
-    ) -> Option<Box<ChainInfo>> {
-        let mut info = None;
-        for certificate in certificates {
-            let hash = certificate.hash();
-            if !certificate.value().is_confirmed() || certificate.value().chain_id() != chain_id {
-                // The certificate is not as expected. Give up.
-                warn!("Failed to process network certificate {}", hash);
-                return info;
-            }
-            let mut result = self
-                .handle_certificate(certificate.clone(), vec![], notifier)
-                .await;
-
-            result = match &result {
-                Err(err) => {
-                    if let Some(blob_ids) = err.get_blobs_not_found() {
-                        let blobs = remote_node.try_download_blobs(blob_ids.as_slice()).await;
-                        if blobs.len() != blob_ids.len() {
-                            result
-                        } else {
-                            self.handle_certificate(certificate, blobs, notifier).await
-                        }
-                    } else {
-                        result
-                    }
-                }
-                _ => result,
-            };
-
-            match result {
-                Ok(response) => info = Some(response.info),
-                Err(error) => {
-                    // The certificate is not as expected. Give up.
-                    warn!("Failed to process network certificate {}: {}", hash, error);
-                    return info;
-                }
-            };
-        }
-        // Done with all certificates.
-        info
-    }
-
     /// Returns a read-only view of the [`ChainStateView`] of a chain referenced by its
     /// [`ChainId`].
     ///
@@ -366,43 +297,6 @@ where
         Ok(response)
     }
 
-    /// Downloads and processes all certificates up to (excluding) the specified height.
-    #[instrument(level = "trace", skip(self, validators, notifier))]
-    pub async fn download_certificates(
-        &self,
-        validators: &[RemoteNode<impl ValidatorNode>],
-        chain_id: ChainId,
-        target_next_block_height: BlockHeight,
-        notifier: &impl Notifier,
-    ) -> Result<Box<ChainInfo>, LocalNodeError> {
-        // Sequentially try each validator in random order.
-        let mut validators = validators.iter().collect::<Vec<_>>();
-        validators.shuffle(&mut thread_rng());
-        for remote_node in validators {
-            let info = self.local_chain_info(chain_id).await?;
-            if target_next_block_height <= info.next_block_height {
-                return Ok(info);
-            }
-            self.try_download_certificates_from(
-                remote_node,
-                chain_id,
-                info.next_block_height,
-                target_next_block_height,
-                notifier,
-            )
-            .await?;
-        }
-        let info = self.local_chain_info(chain_id).await?;
-        if target_next_block_height <= info.next_block_height {
-            Ok(info)
-        } else {
-            Err(LocalNodeError::CannotDownloadCertificates {
-                chain_id,
-                target_next_block_height,
-            })
-        }
-    }
-
     /// Obtains the certificate containing the specified message.
     #[instrument(level = "trace", skip(self))]
     pub async fn certificate_for(
@@ -423,94 +317,6 @@ where
                 ViewError::not_found("could not find certificate with message {}", message_id)
             })?;
         Ok(certificate)
-    }
-
-    /// Downloads and processes all certificates up to (excluding) the specified height from the
-    /// given validator.
-    #[instrument(level = "trace", skip_all)]
-    async fn try_download_certificates_from(
-        &self,
-        remote_node: &RemoteNode<impl ValidatorNode>,
-        chain_id: ChainId,
-        mut start: BlockHeight,
-        stop: BlockHeight,
-        notifier: &impl Notifier,
-    ) -> Result<(), LocalNodeError> {
-        while start < stop {
-            // TODO(#2045): Analyze network errors instead of guessing the batch size.
-            let limit = u64::from(stop)
-                .checked_sub(u64::from(start))
-                .ok_or(ArithmeticError::Overflow)?
-                .min(1000);
-            let Some(certificates) = remote_node
-                .try_query_certificates_from(chain_id, start, limit)
-                .await?
-            else {
-                break;
-            };
-            let Some(info) = self
-                .try_process_certificates(remote_node, chain_id, certificates, notifier)
-                .await
-            else {
-                break;
-            };
-            assert!(info.next_block_height > start);
-            start = info.next_block_height;
-        }
-        Ok(())
-    }
-
-    #[instrument(level = "trace", skip(validators))]
-    async fn download_blob(
-        validators: &[RemoteNode<impl ValidatorNode>],
-        blob_id: BlobId,
-    ) -> Option<Blob> {
-        // Sequentially try each validator in random order.
-        let mut validators = validators.iter().collect::<Vec<_>>();
-        validators.shuffle(&mut thread_rng());
-        for remote_node in validators {
-            if let Some(blob) = remote_node.try_download_blob(blob_id).await {
-                return Some(blob);
-            }
-        }
-        None
-    }
-
-    #[instrument(level = "trace", skip(validators))]
-    pub async fn download_certificate_for_blob_from_validators_futures(
-        validators: &[RemoteNode<impl ValidatorNode>],
-        blob_id: BlobId,
-    ) -> FuturesUnordered<impl Future<Output = Option<Certificate>> + '_> {
-        let futures = FuturesUnordered::new();
-
-        let mut validators = validators.iter().collect::<Vec<_>>();
-        validators.shuffle(&mut thread_rng());
-        for remote_node in validators {
-            futures.push(async move {
-                remote_node
-                    .download_certificate_for_blob(blob_id)
-                    .await
-                    .ok()
-            });
-        }
-
-        futures
-    }
-
-    #[instrument(level = "trace", skip(nodes))]
-    pub async fn download_blobs(
-        blob_ids: &[BlobId],
-        nodes: &[RemoteNode<impl ValidatorNode>],
-    ) -> Vec<Blob> {
-        future::join_all(
-            blob_ids
-                .iter()
-                .map(|blob_id| Self::download_blob(nodes, *blob_id)),
-        )
-        .await
-        .into_iter()
-        .flatten()
-        .collect::<Vec<_>>()
     }
 
     /// Handles any pending local cross-chain requests.
